Add HTTP/1.1 support for CONNECT-UDP
This commit adds support for using HTTP/1.1 as the transport protocol for
CONNECT-UDP proxying, in addition to the existing HTTP/3 support.
(HTTP/2 support is not included due to the lack of support for extended
CONNECT: https://github.com/golang/go/issues/53208.)

This change also includes client configuration flags to choose between
HTTP/1.1 and HTTP/3.  (The server always listens for both.)